requirements.txt
pandas==2.3.2
numpy==2.3.3
langgraph==0.6.7
langchain==0.3.27
matplotlib==3.10.6
python-dotenv==1.1.1
pydantic==2.11.7
langchain-openai==0.3.33


